refactor: move alert and news hooks into dedicated files for better maintainability
